@@ -1,31 +1,50 @@
 {
   "name": "rahul-an-portfolio",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "dev": "concurrently \"npm run server\" \"npm run client\"",
+    "client": "vite",
+    "server": "node api/server.js",
+    "build": "vite build",
+    "preview": "vite preview",
+    "build:css": "postcss src/assets/tailwind.css -o src/assets/main.css",
+    "watch:css": "postcss src/assets/tailwind.css -o src/assets/main.css"
+  },
   "dependencies": {
     "@reduxjs/toolkit": "^1.4.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "axios": ">=0.28.0",
+    "axios": "^1.8.3",
+    "concurrently": "^9.1.2",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.7",
     "emailjs-com": "^2.6.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-hook-form": "^6.11.5",
+    "express": "^4.21.2",
+    "express-rate-limit": "^7.5.0",
+    "framer-motion": "^4.1.17",
+    "node-cache": "^5.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.0",
+    "react-icons": "^5.5.0",
     "react-redux": "^7.2.2",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
     "react-scroll": "^1.8.1",
-    "redux": "^4.0.5"
+    "react-twitter-embed": "^4.0.4",
+    "redux": "^4.0.5",
+    "styled-components": "^6.1.16",
+    "twitter-api-sdk": "^1.2.1"
   },
-  "scripts": {
-    "start": "npm run watch:css && react-scripts --openssl-legacy-provider start",
-    "build": "npm run build:css && react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "build:css": "postcss src/assets/tailwind.css -o src/assets/main.css",
-    "watch:css": "postcss src/assets/tailwind.css -o src/assets/main.css"
+  "devDependencies": {
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "@vitejs/plugin-react": "^4.2.0",
+    "autoprefixer": "^9.8.6",
+    "postcss": "^8.4.31",
+    "postcss-cli": "^7.1.2",
+    "tailwindcss": "^3.3.5",
+    "vite": "^5.0.0"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -42,9 +61,5 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "autoprefixer": "^9.8.6",
-    "postcss-cli": "^7.1.2",
-    "tailwindcss": "^1.8.3"
-  }
+  "proxy": "http://127.0.0.1:3001"
 }
