@@ -1,7 +1,6 @@
-const tailwindcss = require('tailwindcss');
-module.exports = {
-    plugins: [
-        tailwindcss('./tailwind.js'),
-        require('autoprefixer')
-    ],
+export default {
+    plugins: {
+        tailwindcss: {},
+        autoprefixer: {},
+    },
 };
