@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react'
 import Project from './project';
 import {ProjectsList} from './ProjectsList';
+import SocialFeed from '../components/SocialFeed';
 
 export default function Projects() {
     return (
@@ -20,6 +21,6 @@ export default function Projects() {
                     }
                 </div>
             </div>
-        </Fragment>
+                    </Fragment>
     )
 }
